+++ test/object.spec.ts
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import { objToQueryString } from '../dist';
+import { objToQueryString, objToPrettyStr } from '../dist';
 
 describe('Misc', () => {
 	it('funcTimes', (done) => {
@@ -25,6 +25,15 @@ describe('Misc', () => {
 				a: '가나다'
 			}),
 			'a=%EA%B0%80%EB%82%98%EB%8B%A4'
+		);
+		done();
+	});
+
+	it('objToPrettyStr', (done) => {
+		assert.deepStrictEqual(objToPrettyStr({}), '{}');
+		assert.deepStrictEqual(
+			objToPrettyStr({ a: 1, b: { c: 1, d: 2 } }),
+			'{\n\t"a": 1,\n\t"b": {\n\t\t"c": 1,\n\t\t"d": 2\n\t}\n}'
 		);
 		done();
 	});